adds a base class to set up position/color shaders
