+++ libs/webgl/src/programs/LineDrawingProgram.ts
@@ -1,34 +1,12 @@
-import { WebGLProgramBuilder } from "@libs/shared"
-import { Color, BaseProgram } from "@libs/shared"
-import sourceMap, { uniformNames, attributeNames } from "./shaders/sourceMap"
+import { SimpleShaderProgram } from "./abstract/SimpleShaderProgram"
+import { Color } from "@libs/shared"
 
-const VERTEX_SHADER = "position.vertex"
-const FRAGMENT_SHADER = "color.fragment"
-const UNIFORM_NAMES = { ...uniformNames[VERTEX_SHADER], ...uniformNames[FRAGMENT_SHADER] } as const
-const ATTRIBUTE_NAMES = { ...attributeNames[VERTEX_SHADER] } as const
-
-export class LineDrawingProgram extends BaseProgram<keyof typeof UNIFORM_NAMES, keyof typeof ATTRIBUTE_NAMES> {
+export class LineDrawingProgram extends SimpleShaderProgram {
   constructor(
     gl: WebGLRenderingContext,
     public pixelDensity = 1,
   ) {
-    super(LineDrawingProgram.createContextStatic(gl, LineDrawingProgram.createProgramStatic(gl)))
-  }
-
-  protected static createProgramStatic(gl: WebGLRenderingContext): WebGLProgram {
-    return WebGLProgramBuilder.createFromSourceMap(gl, sourceMap, VERTEX_SHADER, FRAGMENT_SHADER)
-  }
-
-  protected createProgram(gl: WebGLRenderingContext): WebGLProgram {
-    return LineDrawingProgram.createProgramStatic(gl)
-  }
-
-  protected static createContextStatic(context: WebGLRenderingContext, program: WebGLProgram) {
-    return BaseProgram.makeBaseContextFromAttributes(context, program, UNIFORM_NAMES, ATTRIBUTE_NAMES)
-  }
-
-  protected createContext(context: WebGLRenderingContext, program: WebGLProgram) {
-    return LineDrawingProgram.createContextStatic(context, program)
+    super(gl)
   }
 
   public syncCanvasSize(): typeof this {
